Refactored sending POST to AlertManager into specific class
